Reversed order of check for RVM non-binary
If you provision once with a failure but past this step, then provision again, like 71 will fail. Reversing the order of the when clause succeeds because item.rc will not equal zero but install_rubies results will be empty and skipped because the first part of the and check fails so the second statement doesnt evaluate.