Add shell wrapper for Vagrant
The goal here is to make launching the vagrant version the same
style as deploying a web version. This parallelism makes it easier
to write instructions and easier to create automation. It also
links the idea that provisioning for Vagrant is parallel to provisioning
the server, while deploying the code is parallel to just turning the
vagrant image up.